Refactor rfxtrx component
